project1/requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.0.7
django-import-export==4.1.1
django-widget-tweaks==1.5.0
Faker==28.4.1
numpy==2.1.1
pandas==2.2.2
pillow==10.4.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
tzdata==2024.1
